Fix for issue - The power ups are not being generated / falling
I fixed the logic on how the power up counter was adding and subtracting the integer value in it. First of all, an if statement was inserted so that if for some reason the counter is less than 0, it is set again to 0. Then the counter increases by 1 only if the counter is 0 or 1 so that the maximum number that can be reached is 1 or 2. When a normal block is hit, the counter decreases only if it is 1 because if it is 0 and it is decreased, it will turn back to 0 and if the counter 2 a power up is generated and the counter comes back to 0.